add: injentando dependencias e corrigindo local de API
