+++ TheMovie/TheMovie/Scenes/Login/MainScreen/Model/MainModel.swift
@@ -11,8 +11,3 @@ struct User {
     let username: String
     let email: String
 }
-
-enum LoginError: Error {
-        case cancelled
-        case noAccessToken
-    }
