+++ TheMovie/TheMovie/Scenes/Login/MainScreen/ViewModel/MainViewModel.swift
@@ -7,22 +7,19 @@
 
 import Foundation
 import GoogleSignIn
-import FBSDKLoginKit
+
+protocol AuthService {
+    func signIn(withPresenting viewController: UIViewController, completion: @escaping (User?, Error?) -> Void)
+}
 
 class MainViewModel {
+    private let authService: AuthService
+    
+    init(authService: AuthService) {
+        self.authService = authService
+    }
+    
     func loginWithGoogle(presentingViewController: UIViewController, completion: @escaping (User?, Error?) -> Void) {
-        GIDSignIn.sharedInstance.signIn(withPresenting: presentingViewController) { signInResult, error in
-            if let error = error {
-                completion(nil, error)
-                return
-            }
-            guard let signInResult = signInResult else {
-                completion(nil, nil)
-                return
-            }
-            let user = signInResult.user
-            let currentUser = User(username: user.profile?.name ?? "", email: user.profile?.email ?? "")
-            completion(currentUser, nil)
-        }
+        authService.signIn(withPresenting: presentingViewController, completion: completion)
     }
 }
